+++ core/database/src/commonMain/kotlin/com/divinelink/core/database/credits/mapper/SeriesCastMapper.kt
@@ -6,7 +6,7 @@ import com.divinelink.core.model.media.MediaItem
 import com.divinelink.core.model.person.Gender
 
 fun FetchShowCast.toPerson(roles: List<PersonRole>): MediaItem.Person = MediaItem.Person(
-  id = id.toInt(),
+  id = id,
   name = name,
   profilePath = profilePath,
   knownForDepartment = knownForDepartment,
