+++ core/model/src/commonMain/kotlin/com/divinelink/core/model/jellyseerr/media/JellyseerrMediaInfo.kt
@@ -1,7 +1,7 @@
 package com.divinelink.core.model.jellyseerr.media
 
 data class JellyseerrMediaInfo(
-  val mediaId: Int,
+  val mediaId: Long,
   val status: JellyseerrStatus.Media,
   val requests: List<JellyseerrRequest>,
   val seasons: List<SeasonRequest>,
