+++ core/model/src/commonMain/kotlin/com/divinelink/core/model/media/MediaItem.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 sealed class MediaItem {
   abstract val uniqueIdentifier: String
-  abstract val id: Int
+  abstract val id: Long
   abstract val name: String
   abstract val posterPath: String?
   abstract val backdropPath: String?
@@ -15,7 +15,7 @@ sealed class MediaItem {
 
   @Serializable
   sealed class Media(override val mediaType: MediaType) : MediaItem() {
-    abstract override val id: Int
+    abstract override val id: Long
     abstract override val name: String
     abstract override val posterPath: String?
     abstract override val backdropPath: String?
@@ -31,7 +31,7 @@ sealed class MediaItem {
 
     @Serializable
     data class TV(
-      override val id: Int,
+      override val id: Long,
       override val name: String,
       override val posterPath: String?,
       override val backdropPath: String?,
@@ -48,7 +48,7 @@ sealed class MediaItem {
 
     @Serializable
     data class Movie(
-      override val id: Int,
+      override val id: Long,
       override val name: String,
       override val posterPath: String?,
       override val backdropPath: String?,
@@ -69,7 +69,7 @@ sealed class MediaItem {
 
   @Serializable
   data class Person(
-    override val id: Int,
+    override val id: Long,
     override val name: String,
     val profilePath: String?,
     val gender: Gender,
@@ -83,7 +83,7 @@ sealed class MediaItem {
   }
 
   data object Unknown : MediaItem() {
-    override val id: Int = -1
+    override val id: Long = -1
     override val name: String = ""
     override val posterPath: String? = null
     override val backdropPath: String? = null
