+++ core/network/src/commonMain/kotlin/com/divinelink/core/network/list/model/MediaItemRequest.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class MediaItemRequest(
-  @SerialName("media_id") val mediaId: Int,
+  @SerialName("media_id") val mediaId: Long,
   @SerialName("media_type") val mediaType: String,
 )
