+++ core/network/src/commonMain/kotlin/com/divinelink/core/network/media/model/details/watchlist/AddToAccountRequestBody.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class AddToAccountRequestBody(
   @SerialName("media_type") val mediaType: String,
-  @SerialName("media_id") val mediaId: Int,
+  @SerialName("media_id") val mediaId: Long,
   val watchlist: Boolean?,
   val favorite: Boolean?,
 )
